admin/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
httpx
python-multipart>=0.0.9
requests
Jinja2

